requirements.txt
streamlit
diffusers
torch
transformers
accelerate
pillow
python-dotenv 